feat: rename speaker layout to intro
